+++ src/dataAccess/DeleteFailedException.java
@@ -1,7 +1,7 @@
 package dataAccess;
 
 public class DeleteFailedException extends DAOOperationFailedException {
-    public DeleteFailedException(String objectType, int objectId, String message) {
+    public DeleteFailedException(String objectType, long objectId, String message) {
         super("deletion", objectType, objectId, message);
     }
 }
