+++ src/dataAccess/InsertionFailedException.java
@@ -1,7 +1,7 @@
 package dataAccess;
 
 public class InsertionFailedException extends DAOOperationFailedException {
-    public InsertionFailedException(String objectType, int objectId, String message) {
+    public InsertionFailedException(String objectType, long objectId, String message) {
         super("insertion", objectType, objectId, message);
     }
 }
